% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dijkstraAlgorithm}
\alias{dijkstraAlgorithm}
\title{dijkstra algorithm}
\usage{
dijkstraAlgorithm(adj_matrix, starting)
}
\arguments{
\item{adj_matrix}{the adjacency matrix}

\item{starting}{the starting vertex we want to run the algorithm from}
}
\value{
a 2*n matrix (first row : the best distance from the starting vertex vector , second row : the previous vertex vector)
}
\description{
dijkstra algorithm
}
